Added mutex in shortlink generation to allow burst requests
In case of multiple simultaneous requests, the code starts to generate
same shortlink for multiple long URLs and fails when it tries to insert
the shortlinks into database. To avoid this, only one process at a time
should try link generation while others should wait until it finishes.
To avoid long waiting times, the process will fail if it has to wait
for more than 3000 miliseconds to acquire the lock.